+++ src/main/java/com/github/mjkuranda/spaceadventure2/SpaceAdventure2.java
@@ -1,5 +1,7 @@
 package com.github.mjkuranda.spaceadventure2;
 
+import com.github.mjkuranda.spaceadventure2.resources.GameFont;
+import com.github.mjkuranda.spaceadventure2.resources.GameImage;
 import com.github.mjkuranda.spaceadventure2.states.GameOverState;
 import com.github.mjkuranda.spaceadventure2.states.GameState;
 import com.github.mjkuranda.spaceadventure2.states.IntroState;
@@ -8,6 +10,7 @@ import org.newdawn.slick.*;
 import org.newdawn.slick.state.StateBasedGame;
 
 import java.awt.*;
+import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
@@ -21,13 +24,18 @@ public class SpaceAdventure2 extends StateBasedGame {
 
     @Override
     public void initStatesList(GameContainer container) throws SlickException {
+        // FIXME: Loading resources in `LoadingState` state.
+        GameImage.load();
+        try {
+            GameFont.load();
+        } catch (IOException | FontFormatException e) {
+            throw new RuntimeException(e);
+        }
+
         addState(new MainMenuState(this));
         addState(new GameState());
         addState(new GameOverState());
         addState(new IntroState());
-
-        GameImages images = GameImages.getInstance();
-        images.init();
     }
 
     public static void main(String[] args) {