+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/Asteroid.java
@@ -1,13 +1,14 @@
 package com.github.mjkuranda.spaceadventure2.entities;
 
 import com.github.mjkuranda.spaceadventure2.entities.shootable.CantShoot;
+import com.github.mjkuranda.spaceadventure2.resources.GameImage;
 
 import java.util.LinkedList;
 
 public class Asteroid extends SpaceEntity {
 
     public Asteroid(LinkedList<Entity> subscriber, float x, float y) {
-        super(EntityType.ASTEROID, subscriber, EntityTurn.INCOMING, new CantShoot(), EntityShapes.ASTEROID_SHAPE, x, y, 0.02f, 100, 1);
+        super(EntityType.ASTEROID, subscriber, EntityTurn.INCOMING, new CantShoot(), EntityShapes.ASTEROID_SHAPE, GameImage.ASTEROID, x, y, 0.02f, 100, 1);
     }
 
     public Asteroid(float x, float y) {