+++ src/main/java/com/github/mjkuranda/spaceadventure2/entities/SpaceEntity.java
@@ -1,7 +1,7 @@
 package com.github.mjkuranda.spaceadventure2.entities;
 
-import com.github.mjkuranda.spaceadventure2.entities.missiles.Missile;
 import com.github.mjkuranda.spaceadventure2.entities.shootable.Shootable;
+import org.newdawn.slick.Image;
 import org.newdawn.slick.geom.Shape;
 
 import java.util.LinkedList;
@@ -11,8 +11,8 @@ public abstract class SpaceEntity extends Entity {
     /** Shootable strategy */
     private Shootable shootable;
 
-    public SpaceEntity(EntityType type, LinkedList<Entity> subscriber, EntityTurn turn, Shootable shootable, Shape shape, float x, float y, float speed, int durability, int score) {
-        super(type, subscriber, turn, shape, x, y, speed, durability, score);
+    public SpaceEntity(EntityType type, LinkedList<Entity> subscriber, EntityTurn turn, Shootable shootable, Shape shape, Image image, float x, float y, float speed, int durability, int score) {
+        super(type, subscriber, turn, shape, image, x, y, speed, durability, score);
 
         this.shootable = shootable;
     }